Fixes to STAMP Benchmarks
* genome.Genome => one of the worker threads was running in the main thread,
which was replaced by a new thread similar to the others. Furthermore, the
await() condition through the genome.Barrier object was replaced by the join()
to each worker thread.

* genome.Sequencer => the empty version of the atomicMethodFive was replaced by
the atomicMethodFive2 in comments, which includes an implementation. With the
original  empty atomicMethodFive the Genome was crashing.

* kMeans .kMeans => replace the FileImputStream with getResourceAsStream that
allows to read the configuration file from any location in the classpath. In
this case, we are reading the KMeans input file from the deuceTests.jar.

* ssca2.GetUserParameters => change the type of two arguments from double to integer.

* vacation.Client => move the parameterization arrays into their own atomic
operation. This change turns these arrays private to their transactions.

+++ src/test/jstamp/ssca2/GetUserParameters.java
@@ -109,7 +109,7 @@ public class GetUserParameters {
         //check options
         if(arg.equals("-i")) {
           if(i < args.length) {
-            glb.PROB_INTERCL_EDGES = new Integer(args[i++]).floatValue();
+            glb.PROB_INTERCL_EDGES = new Double(args[i++]).floatValue();
           }
         } else if(arg.equals("-k")) {
           if(i < args.length) {
@@ -137,7 +137,7 @@ public class GetUserParameters {
           }
         } else if(arg.equals("-u")) {
           if(i < args.length) {
-            glb.PROB_UNIDIRECTIONAL = new Integer(args[i++]).floatValue();
+            glb.PROB_UNIDIRECTIONAL = new Double(args[i++]).floatValue();
           }
         } else if(arg.equals("-w")) {
           if(i < args.length) {
